Renamed files to make more sense. Created checklist for group members
