+++ app/CreateGroup.js
@@ -1,10 +1,38 @@
 import React, { Component } from 'react';
-import { View, Text, StyleSheet, TextInput } from 'react-native';
+import CheckBox from 'react-native-checkbox';
+import { endpoint } from './endpoint.js';
+import { View, Text, StyleSheet, TextInput, ListView } from 'react-native';
 
 export default class CreateGroup extends Component {
   constructor(props) {
     super(props);
-    this.state =  { text: '' };
+    this.state =  { 
+      friends: new ListView.DataSource({
+        rowHasChanged: (row1, row2) => row1 !== row2
+      }),
+      text: '',
+    };
+  }
+
+  componentWillMount() {
+    console.log("in component will mount in all friends list")
+    fetch(`${endpoint}/api/friends`, {
+      method: 'GET',
+    })
+      .then(function(response) {
+        return response.json();
+      })
+      .then((friends) => {
+        console.log(friends);
+        this.setState({
+          friends: this.state.friends.cloneWithRows(friends),
+          loaded: true
+        });
+      })
+      .catch((error) => {
+        console.log('There was an error in fetching your data: ', error);
+        return error;
+      });
   }
 
   static navigationOptions = {
@@ -21,6 +49,14 @@ export default class CreateGroup extends Component {
           onChangeText={(text) => this.setState({text})}
           placeholder='Insert Group Name'
           value={this.state.text}
+        />
+        <ListView
+          dataSource={this.state.friends}
+          renderRow={(rowData) => (
+            <CheckBox
+              label={rowData.first}
+            />
+          )}
         />
       </View>
     );
