+++ app/index.js
@@ -4,7 +4,7 @@ import AddFence from './AddFence.js';
 import GroupsList from './GroupsList.js';
 import AllFriendsList from './AllFriendsList.js';
 import Settings from './Settings.js';
-import AddFriend from './AddFriend.js';
+import AddFriendList from './AddFriendList.js';
 import CreateGroup from './CreateGroup.js';
 
 import React from 'react';
@@ -17,24 +17,24 @@ const HomePageTabs = TabNavigator({
   AllFriendsList: { screen: AllFriendsList },
 });
 
-// HomePageTabs.navigationOptions = {
-//   title: 'Safely',
-//   header: (({ navigate }) => ({
-//     left: <Button title='Menu' onPress={ () => navigate('DrawerOpen') }/>
-//   }))
-// };
-
-HomePageTabs.navigationOptions = ({navigation}) => ({
+HomePageTabs.navigationOptions = {
   title: 'Safely',
-  headerRight: <Button title='Menu' onPress={ () => navigation.navigate('DrawerOpen') } />
-});
+  header: (({ navigate }) => ({
+    left: <Button title='Menu' onPress={ () => navigate('DrawerOpen') }/>
+  }))
+};
+
+// HomePageTabs.navigationOptions = ({navigation}) => ({
+//   title: 'Safely',
+//   headerRight: <Button title='Menu' onPress={ () => navigation.navigate('DrawerOpen') } />
+// });
 
 const Stack = StackNavigator({
   HomePageTabs: { screen: HomePageTabs },
   FriendMap: { screen: FriendMap },
   AddFence: { screen: AddFence },
   Settings: { screen: Settings },
-  AddFriend: { screen: AddFriend },
+  AddFriend: { screen: AddFriendList },
   CreateGroup: { screen: CreateGroup },
 });
 
@@ -44,7 +44,7 @@ Stack.navigationOptions = {
 
 const MenuDrawer = DrawerNavigator({
   Stack: { screen: Stack },
-  AddFriend: { screen: AddFriend },  
+  AddFriend: { screen: AddFriendList },  
   Settings: { screen: Settings }
 
 }, {
