+++ package.json
@@ -9,6 +9,7 @@
   "dependencies": {
     "react": "16.0.0-alpha.6",
     "react-native": "0.43.2",
+    "react-native-checkbox": "^1.1.0",
     "react-native-contacts": "^0.8.1",
     "react-native-fcm": "^6.2.1",
     "react-native-geo-fencing": "^0.1.0",
